feat(global-setup): add environment and API validation during global setup - once per run
